Continuing merge of tasks/ modules
