+++ .gitmodules
@@ -4,3 +4,6 @@
 [submodule "third-party/kernels"]
 	path = third-party/kernels
 	url = git@github.com:faasm/Kernels.git
+[submodule "third-party/kernels-native"]
+	path = third-party/kernels-native
+	url = https://github.com/faasm/Kernels
