+++ docker-compose.yml
@@ -1,7 +1,7 @@
 version: "3"
 
 services:
-  cli:
+  lammps-cli:
     image: ${LAMMPS_CLI_IMAGE}
     working_dir: /code/experiment-lammps
     command: ${INNER_SHELL}
@@ -10,3 +10,13 @@ services:
     network_mode: host
     volumes:
       - .:/code/experiment-lammps
+
+  kernels-cli:
+    image: ${KERNELS_CLI_IMAGE}
+    working_dir: /code/experiment-kernels
+    command: ${INNER_SHELL}
+    stdin_open: true
+    tty: true
+    network_mode: host
+    volumes:
+      - .:/code/experiment-kernels
