+++ tasks/__init__.py
@@ -1,15 +1,15 @@
 from invoke import Collection
 
-from . import container
-from . import data
 from . import native
-from . import run
-from . import wasm
+from . import openmpi
+
+from tasks.kernels import ns as kernels_ns
+from tasks.lammps import ns as lammps_ns
 
 ns = Collection(
-    container,
-    data,
     native,
-    run,
-    wasm,
+    openmpi,
 )
+
+ns.add_collection(lammps_ns, name="lammps")
+ns.add_collection(kernels_ns, name="kernels")
