+++ tasks/kernels/__init__.py
@@ -1,0 +1,6 @@
+from invoke import Collection
+
+from . import build
+from . import container
+
+ns = Collection(build, container)
