+++ tasks/kernels/build.py
@@ -3,7 +3,7 @@ from os.path import join
 from subprocess import run
 from sys import exit
 
-from tasks.util import NATIVE_CODE_DIR, WASM_CODE_DIR, WASM_DIR, FAASM_DIR
+from tasks.util import PROJ_ROOT, WASM_DIR, FAASM_DIR
 
 MAKE_TARGETS = [
     ("MPI1/Synch_global", "global"),
@@ -42,7 +42,9 @@ def build(ctx, mode, clean=False):
 
         if res.returncode != 0:
             print(
-                "Making kernel in {} with target {} failed.".format(subdir, make_target)
+                "Making kernel in {} with target {} failed.".format(
+                    subdir, make_target
+                )
             )
             return
 
@@ -54,6 +56,8 @@ def upload_wasm(ctx):
     """
     for target in [t[1] for t in MAKE_TARGETS]:
         wasm_src = join(WASM_DIR, "{}.wasm".format(target))
-        cmd = "inv -r faasmcli/faasmcli upload prk {} {}".format(target, wasm_src)
+        cmd = "inv -r faasmcli/faasmcli upload prk {} {}".format(
+            target, wasm_src
+        )
         print(cmd)
         run(cmd, shell=True, check=True, cwd=FAASM_DIR)
