+++ tasks/native.py
@@ -1,11 +1,6 @@
 from invoke import task
-from os.path import join
 from tasks.util import (
-    NATIVE_BUILD_DIR,
-    LAMMPS_DIR,
-    NATIVE_INSTALL_DIR,
     NATIVE_HOSTFILE,
-    clean_dir,
     run_kubectl_cmd,
     get_pod_names_ips,
 )
@@ -13,38 +8,6 @@ from subprocess import run
 
 HOSTFILE_LOCAL_FILE = "/tmp/hostfile"
 SLOTS_PER_HOST = 2
-
-
-@task(default=True)
-def build(ctx, clean=False, verbose=False):
-    """
-    Build and install LAMMPS natively
-    """
-    # The LAMMPS CMake build instructions can be found in the following link
-    # https://lammps.sandia.gov/doc/Build_cmake.html
-
-    cmake_dir = join(LAMMPS_DIR, "cmake")
-
-    clean_dir(NATIVE_BUILD_DIR, clean)
-    clean_dir(NATIVE_INSTALL_DIR, clean)
-
-    cmake_cmd = [
-        "cmake",
-        "-GNinja",
-        "-DCMAKE_C_COMPILER=/usr/bin/clang-10",
-        "-DCMAKE_CXX_COMPILER=/usr/bin/clang++-10",
-        "-DCMAKE_INSTALL_PREFIX={}".format(NATIVE_INSTALL_DIR),
-        cmake_dir,
-    ]
-
-    cmake_str = " ".join(cmake_cmd)
-    print(cmake_str)
-
-    run(cmake_str, check=True, shell=True, cwd=NATIVE_BUILD_DIR)
-
-    run("ninja", check=True, shell=True, cwd=NATIVE_BUILD_DIR)
-
-    run("ninja install", check=True, shell=True, cwd=NATIVE_BUILD_DIR)
 
 
 @task
